Split template utils into their own file
