+++ src/utils/workspaceUtils.ts
@@ -6,14 +6,11 @@
 import * as path from 'path';
 import { commands, MessageItem, OpenDialogOptions, Uri, window, workspace, WorkspaceFolder } from "vscode";
 import { IActionContext, IAzureQuickPickItem, UserCancelledError } from "vscode-azureextensionui";
-import { RepoNameStep } from '../commands/createRepo/RepoNameStep';
-import { GitHubOrgListStep } from '../commands/createStaticWebApp/GitHubOrgListStep';
-import { IStaticWebAppWizardContext } from '../commands/createStaticWebApp/IStaticWebAppWizardContext';
 import { cloneRepo } from '../commands/github/cloneRepo';
 import { NoWorkspaceError } from '../errors';
-import { RepoData } from '../gitHubTypings';
-import { createRepoFromTemplate, getTemplateRepos } from './gitHubUtils';
+import { createRepoFromTemplate } from './gitHubUtils';
 import { localize } from "./localize";
+import { pickTemplate, promptForTemplateRepoName } from './templateUtils';
 
 export function getSingleRootFsPath(): string | undefined {
     // if this is no workspace or a multi-root workspace, return undefined
@@ -104,33 +101,4 @@ export async function getWorkspaceFolder(context: IActionContext): Promise<Works
     }
 
     return folder;
-}
-
-async function pickTemplate(context: IActionContext): Promise<RepoData> {
-    const templateRepos: RepoData[] = await getTemplateRepos(context);
-
-    const placeHolder: string = localize('chooseTemplatePrompt', 'Choose a static web app template for the new repository.');
-    const picks: IAzureQuickPickItem<RepoData>[] = templateRepos.map((repo) => { return { label: repo.name, data: repo }; });
-    const pick: IAzureQuickPickItem<RepoData> = await context.ui.showQuickPick(picks, { placeHolder, suppressPersistence: true });
-
-    return pick.data;
-}
-
-async function promptForTemplateRepoName(context: IActionContext, template: RepoData): Promise<string> {
-    const templateName = template.name;
-
-    const validateNameContext: Partial<IStaticWebAppWizardContext> & IActionContext = {
-        ...context,
-        orgData: await GitHubOrgListStep.getAuthenticatedUser(context)
-    }
-
-    const validateRepoName = async (value: string): Promise<string | undefined> => await RepoNameStep.validateRepoName(validateNameContext, value);
-    const isTemplateNameAValidRepoName = !(await validateRepoName(templateName));
-    const newRepoName = await context.ui.showInputBox({
-        validateInput: validateRepoName,
-        prompt: localize('newRepoFromTemplatePrompt', 'Enter the name of the new GitHub repository to create from the "{0}" template.', templateName),
-        value: isTemplateNameAValidRepoName ? templateName : ''
-    });
-
-    return newRepoName;
 }
